Add: `decorate.py`のテストファイルを追加
